# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "\"赚圈圈API 返回的goods如果其中的goodsin为yes代表它可以进行交易，即：未下架\"",
        "title": "miniproject",
        "termsOfService": "http://swagger.io/terrms/",
        "contact": {
            "name": "yyj",
            "email": "2105753640@qq.com"
        },
        "version": "1.0.0"
    },
    "host": "119.3.133.235:8080",
    "basePath": "/api/vi",
    "paths": {
...
        "/money/search": {
            "post": {
                "description": "\"order=1-\u003e返回前十个商品的内容，summary不需要展示出来-》在商品详情页里,搜索的api\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Good"
                ],
                "summary": "\"搜索并返回已排序的商品信息\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜索框输入的内容",
                        "name": "content",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "msg\":\"search successfully\",\"infor\":[]tables.Good}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "204": {
                        "description": "msg\":\"find nothing\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
...
}